services/ocr/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
Pillow==10.4.0
pytesseract==0.3.10
PyMuPDF==1.24.8


